+++ src/sqlite/init.rs
@@ -13,7 +13,7 @@
 // You should have received a copy of the GNU Affero General Public License
 // along with this program.  If not, see <http://www.gnu.org/licenses/>.
 
-use crate::logger::log;
+use crate::prelude::*;
 use crate::resource::Resource;
 use rusqlite::params;
 
@@ -24,18 +24,18 @@ pub fn init_db(db: &mut rusqlite::Connection, prefix: &str) -> Result<(), rusqli
             Ok(version)
         })
         .unwrap_or(-1);
+    info!("Found event database schema version {}", version);
     let mut next_version = version + 1;
 
     loop {
         let filename = format!("{}/V{}.sql", prefix, next_version);
         if let Some(asset) = Resource::get(&filename) {
             if next_version == 0 {
-                log::info!("Initializing SQLite database ({})", prefix)
+                info!("Initializing SQLite database ({})", prefix)
             } else {
-                log::info!(
+                info!(
                     "Updating SQLite database to schema version {} ({})",
-                    next_version,
-                    prefix
+                    next_version, prefix
                 );
             }
             let asset = String::from_utf8_lossy(&asset);
@@ -48,6 +48,10 @@ pub fn init_db(db: &mut rusqlite::Connection, prefix: &str) -> Result<(), rusqli
             tx.commit()?;
             next_version += 1;
         } else {
+            debug!(
+                "Did not find resource file {}, database migration done",
+                filename
+            );
             break;
         }
     }